+++ src/client.js
@@ -1,8 +1,9 @@
 const snekfetch = require('snekfetch');
 const Package = require('../package.json');
 
-const Profile = require('./profile.js');
-const Account = require('./account.js');
+const Profile = require('./Profile');
+const Account = require('./Account');
+const Match = require('./Match');
 
 class Client {
     constructor(key) {
@@ -19,22 +20,29 @@ class Client {
         };
     }
 
-    getProfile(username) {
-        return this._apiRequest(`https://pubgtracker.com/api/profile/pc/${username}`)
-            .then(body => new Profile(body))
+    getProfile(username, options = {}) {
+        return this._apiRequest(`https://api.pubgtracker.com/v2/profile/pc/${username}`, options)
+            .then(body => new Profile(body, this))
             .catch(e => Promise.reject(e));
     }
 
     getAccount(id) {
-        return this._apiRequest(`https://pubgtracker.com/api/search?steamId=${id}`)
-            .then(body => new Account(body))
+        return this._apiRequest(`https://api.pubgtracker.com/v2/search/steam?steamId=${id}`)
+            .then(body => new Account(body, this))
             .catch(e => Promise.reject(e));
     }
 
-    _apiRequest(url) {
+    getMatchHistory(accountId) {
+        return this._apiRequest(`https://api.pubgtracker.com/v2/matches/pc/${accountId}`)
+            .then(body => body.map(match => new Match(match)))
+            .catch(e => Promise.reject(e));
+    }
+
+    _apiRequest(url, options = {}) {
         return snekfetch.get(url)
             .set(this._headers)
-            .then(r => r.body.error ? Promise.reject(r.body.message || r.body.error) : r.body)
+            .query(options)
+            .then(r => r.body.error ? Promise.reject(`"${r.body.error}" with code ${r.body.code}`) : r.body)
             .catch(e => Promise.reject(`HTTP ${e}`));
     }